# openapi.yml
openapi: 3.1.0
info:
  title: Masashi Enokida API
  version: 1.0.0
servers:
  - url: http://localhost:4000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    DiscographyItem:
      type: object
      required: [id, title, release_year, sort_order]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        release_year:
          type: integer
        description:
          type: string
          nullable: true
        image_url:
          type: string
          nullable: true
        purchase_url:
          type: string
          nullable: true
        sort_order:
          type: integer
    DiscographyInput:
      type: object
      required: [title, release_year]
      properties:
        title:
          type: string
        release_year:
          type: integer
        description:
          type: string
          nullable: true
        image_url:
          type: string
          nullable: true
        purchase_url:
          type: string
          nullable: true
        sort_order:
          type: integer
          default: 0
        is_published:
          type: boolean
          default: true
...
    AdminDiscographyItem:
      type: object
      required: [id, title, release_year, sort_order, is_published, created_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        release_year:
          type: integer
        description:
          type: string
          nullable: true
        image_url:
          type: string
          nullable: true
        purchase_url:
          type: string
          nullable: true
        sort_order:
          type: integer
        is_published:
          type: boolean
        created_at:
          type: string
          format: date-time
...
paths:
...
  /discography:
    get:
      operationId: getDiscography
      summary: List discography
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiscographyItem'
    post:
      operationId: createDiscography
      summary: Create discography item (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscographyInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscographyItem'
  /discography/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getDiscographyById
      summary: Get single discography item
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscographyItem'
        '404':
          description: Not found
    put:
      operationId: updateDiscography
      summary: Update discography item (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscographyInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscographyItem'
    delete:
      operationId: deleteDiscography
      summary: Delete discography item (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
...
  /admin/discography:
    get:
      operationId: getAdminDiscography
      summary: List all discography including unpublished (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminDiscographyItem'
  /admin/discography/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getAdminDiscographyById
      summary: Get single discography item (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDiscographyItem'
        '404':
          description: Not found
...